refactor(robo): moved data loading into robo.start() in production mode
This allows for simpler direct usage without needing to know the internals for loading this externally. Ya know, like when migrating existing projects.

State and Flashcore now load automatically, even when no `stateLoad` promise has been passed.
