+++ .changeset/chatty-crews-bake.md
@@ -1,0 +1,5 @@
+---
+'robo.js': patch
+---
+
+refactor(core): moved data loading into robo.start() in production mode
