+++ packages/robo/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "robo.js",
-	"version": "0.10.9",
+	"version": "0.10.9-dev",
 	"private": false,
 	"description": "Power up Discord with effortless activities, bots, web servers, and more!",
 	"keywords": [
